Move inlined functions
